refactor: Eradicate God Objects into domain-specific modules (Phase 2.95 Step 0)
This commit is a pure structural refactor containing NO changes to underlying business logic. It establishes the architectural foundation required for the upcoming OpenTherm CQRS cut-over (Phase 2.95, Step F5).

- Renamed `device/` directory to `devices/` to adhere to collection naming conventions.
- Split `heat.py` (~1,100 lines), `hvac.py` (~600 lines), and `models.py` into flat, globally uniquely-named, domain-specific modules (e.g., `heat_actuators.py`, `state_climate.py`).
- Created public API facades in `devices/__init__.py`, `models/__init__.py`, and `commands/__init__.py` to maintain backwards compatibility for downstream imports.
- Achieved 100% `mypy --strict` compliance by resolving base class initialisation conflicts, type narrowing, and Liskov Substitution Principle (LSP) override signatures.
- Fixed unit test import paths and corrected `@patch` targeting for `ramses_tx.command.Command`.
- Restored `RfgGateway` class and updated all `pytest` snapshot baselines to reflect the new module paths.

Test suite passing: 33,298 passed, 64 skipped, 0 strict mypy errors.
