+++ src/ramses_rf/commands/__init__.py
@@ -1,0 +1,7 @@
+"""RAMSES RF - Commands package."""
+
+from .core import Command
+
+__all__ = [
+    "Command",
+]
