+++ src/ramses_rf/devices/hvac_ventilators.py
@@ -1,5 +1,4 @@
-#!/usr/bin/env python3
-"""RAMSES RF - devices from the HVAC domain."""
+"""RAMSES RF - HVAC Ventilator & Gateway Devices."""
 
 from __future__ import annotations
 
@@ -8,17 +7,14 @@ import contextlib
 import logging
 from collections.abc import Callable
 from datetime import timedelta as td
-from typing import TYPE_CHECKING, Any, TypeVar
+from typing import TYPE_CHECKING, Any
 
 from ramses_rf import exceptions as exc
-from ramses_rf.address import Address
 from ramses_rf.const import (
     HEARTBEAT_TIMEOUT_FILTER,
-    HEARTBEAT_TIMEOUT_REMOTE,
-    HEARTBEAT_TIMEOUT_SENSOR,
+    RQ,
     SZ_AIR_QUALITY,
     SZ_AIR_QUALITY_BASIS,
-    SZ_BOOST_TIMER,
     SZ_BYPASS_MODE,
     SZ_BYPASS_POSITION,
     SZ_BYPASS_STATE,
@@ -38,7 +34,6 @@ from ramses_rf.const import (
     SZ_OUTDOOR_TEMP,
     SZ_POST_HEAT,
     SZ_PRE_HEAT,
-    SZ_PRESENCE_DETECTED,
     SZ_REMAINING_MINS,
     SZ_REQ_REASON,
     SZ_REQ_SPEED,
@@ -47,28 +42,21 @@ from ramses_rf.const import (
     SZ_SUPPLY_FLOW,
     SZ_SUPPLY_TEMP,
     SZ_TEMPERATURE,
+    Code,
     DevType,
 )
-from ramses_rf.entity import class_by_attr
 from ramses_rf.helpers import schedule_task
-from ramses_rf.models import HvacState
+from ramses_rf.models import DeviceTraits, HvacState
 from ramses_tx import Command, Packet, Priority
 from ramses_tx.typing import PayloadT
 
-from ..messages import Message
-from ..protocol.ramses import CODES_OF_HVAC_DOMAIN_ONLY, HVAC_KLASS_BY_VC_PAIR
-from .base import BatteryState, DeviceHvac, Fakeable
-
-from ramses_rf.const import (  # noqa: F401, isort: skip, pylint: disable=unused-import
-    I_,
-    RP,
-    RQ,
-    W_,
-    Code,
-)
+from .dev_base import DeviceHvac
 
 if TYPE_CHECKING:
-    from ramses_rf.models import DeviceTraits
+    from ramses_rf.devices.dev_base import Device
+    from ramses_rf.systems import Evohome
+
+    from ..messages import Message
 
 # TODO: Switch this module to utilise the (run-time) decorator design pattern...
 # - https://refactoring.guru/design-patterns/decorator/python/example
@@ -83,210 +71,6 @@ if TYPE_CHECKING:
 
 
 _LOGGER = logging.getLogger(__name__)
-
-
-_HvacRemoteBaseT = TypeVar("_HvacRemoteBaseT", bound="HvacRemoteBase")
-_HvacSensorBaseT = TypeVar("_HvacSensorBaseT", bound="HvacSensorBase")
-
-
-class HvacRemoteBase(DeviceHvac):
-    """Base class for HVAC remote control devices.
-
-    This class serves as a base for all remote control devices in the HVAC domain.
-    It provides common functionality and interfaces for remote control operations.
-    """
-
-    def __init__(
-        self, *args: Any, traits: DeviceTraits | None = None, **kwargs: Any
-    ) -> None:
-        """Initialize the HvacRemoteBase class.
-
-        :param args: Positional arguments passed to the parent class
-        :param traits: Strictly typed traits object for device creation
-        :param kwargs: Keyword arguments passed to the parent class
-        """
-        super().__init__(*args, traits=traits, **kwargs)
-        if not hasattr(self, "hvac_state"):
-            self.hvac_state = HvacState()
-
-    def _post_class_promote(self) -> None:
-        """Initialize state when promoted from a generic HVAC device."""
-        if not hasattr(self, "hvac_state"):
-            self.hvac_state = HvacState()
-
-    @property
-    def heartbeat_timeout(self) -> td:
-        """Return the timeout before the device is considered unavailable.
-
-        :return: The timeout duration.
-        :rtype: td
-        """
-        return HEARTBEAT_TIMEOUT_REMOTE
-
-
-class HvacSensorBase(DeviceHvac):
-    """Base class for HVAC sensor devices.
-
-    This class serves as a base for all sensor devices in the HVAC domain.
-    It provides common functionality for sensor data collection and processing.
-    """
-
-    def __init__(
-        self, *args: Any, traits: DeviceTraits | None = None, **kwargs: Any
-    ) -> None:
-        """Initialize the HvacSensorBase class.
-
-        :param args: Positional arguments passed to the parent class
-        :param traits: Strictly typed traits object for device creation
-        :param kwargs: Keyword arguments passed to the parent class
-        """
-        super().__init__(*args, traits=traits, **kwargs)
-        if not hasattr(self, "hvac_state"):
-            self.hvac_state = HvacState()
-
-    def _post_class_promote(self) -> None:
-        """Initialize state when promoted from a generic HVAC device."""
-        if not hasattr(self, "hvac_state"):
-            self.hvac_state = HvacState()
-
-    @property
-    def heartbeat_timeout(self) -> td:
-        """Return the timeout before the device is considered unavailable.
-
-        :return: The timeout duration.
-        :rtype: td
-        """
-        return HEARTBEAT_TIMEOUT_SENSOR
-
-
-class CarbonDioxide(HvacSensorBase):  # 1298
-    """The CO2 sensor (cardinal code is 1298)."""
-
-    async def co2_level(self) -> int | None:
-        """Get the CO2 level in ppm.
-
-        :return: The CO2 level in parts per million (ppm), or None if not available
-        :rtype: int | None
-        """
-        return self.hvac_state.co2_level
-
-    async def set_co2_level(self, value: int | None) -> Packet:
-        """Set a fake CO2 level for the sensor.
-
-        :param value: The CO2 level in ppm to set, or None to clear the fake value
-        :type value: int | None
-        :raises TypeError: If the sensor is not in faked mode
-        :return: The sent packet
-        :rtype: Packet
-        """
-
-        if not self.is_faked:
-            raise exc.DeviceNotFaked(f"{self}: Faking is not enabled")
-
-        cmd = Command.put_co2_level(self.id, value)
-        return await self._gwy.async_send_cmd(
-            cmd, num_repeats=2, priority=Priority.HIGH
-        )
-
-    async def status(self) -> dict[str, Any]:
-        """Return the status of the CO2 sensor.
-
-        :return: A dictionary containing the sensor's status including CO2 level
-        :rtype: dict[str, Any]
-        """
-        base_status = await super().status()
-        return {
-            **base_status,
-            SZ_CO2_LEVEL: await self.co2_level(),
-        }
-
-
-class IndoorHumidity(HvacSensorBase):  # 12A0
-    """The relative humidity sensor (12A0)."""
-
-    async def indoor_humidity(self) -> float | None:
-        """Get the indoor relative humidity.
-
-        :return: The indoor relative humidity as a percentage (0-100), or None if not available
-        :rtype: float | None
-        """
-        return self.hvac_state.indoor_humidity
-
-    async def set_indoor_humidity(self, value: float | None) -> Packet:
-        """Set a fake indoor humidity value for the sensor.
-
-        :param value: The humidity percentage to set (0-100), or None to clear the fake value
-        :type value: float | None
-        :raises TypeError: If the sensor is not in faked mode
-        :return: The sent packet
-        :rtype: Packet
-        """
-
-        if not self.is_faked:
-            raise exc.DeviceNotFaked(f"{self}: Faking is not enabled")
-
-        cmd = Command.put_indoor_humidity(self.id, value)
-        return await self._gwy.async_send_cmd(
-            cmd, num_repeats=2, priority=Priority.HIGH
-        )
-
-    async def status(self) -> dict[str, Any]:
-        """Return the status of the indoor humidity sensor.
-
-        :return: A dictionary containing the sensor's status including humidity level
-        :rtype: dict[str, Any]
-        """
-        base_status = await super().status()
-        return {
-            **base_status,
-            SZ_INDOOR_HUMIDITY: await self.indoor_humidity(),
-        }
-
-
-class PresenceDetect(HvacSensorBase):  # 2E10
-    """The presence sensor (2E10/31E0)."""
-
-    # .I --- 37:154011 --:------ 37:154011 1FC9 030 00-31E0-96599B 00-1298-96599B 00-2E10-96599B 01-10E0-96599B 00-1FC9-96599B    # CO2, idx|10E0 == 01
-    # .W --- 28:126620 37:154011 --:------ 1FC9 012 00-31D9-49EE9C 00-31DA-49EE9C                                                 # FAN, BRDG-02A55
-    # .I --- 37:154011 28:126620 --:------ 1FC9 001 00                                                                            # CO2, incl. integrated control, PIR
-
-    async def presence_detected(self) -> bool | None:
-        """Get the presence detection status.
-
-        :return: True if presence is detected, False if not, None if status is unknown
-        :rtype: bool | None
-        """
-        return self.hvac_state.presence_detected
-
-    async def set_presence_detected(self, value: bool | None) -> Packet:
-        """Set a fake presence detection state for the sensor.
-
-        :param value: The presence state to set (True/False), or None to clear the fake value
-        :type value: bool | None
-        :raises TypeError: If the sensor is not in faked mode
-        :return: The sent packet
-        :rtype: Packet
-        """
-
-        if not self.is_faked:
-            raise exc.DeviceNotFaked(f"{self}: Faking is not enabled")
-
-        cmd = Command.put_presence_detected(self.id, value)
-        return await self._gwy.async_send_cmd(
-            cmd, num_repeats=2, priority=Priority.HIGH
-        )
-
-    async def status(self) -> dict[str, Any]:
-        """Return the status of the presence sensor.
-
-        :return: A dictionary containing the sensor's status including presence detection state
-        :rtype: dict[str, Any]
-        """
-        base_status = await super().status()
-        return {
-            **base_status,
-            SZ_PRESENCE_DETECTED: await self.presence_detected(),
-        }
 
 
 class FilterChange(DeviceHvac):  # FAN: 10D0
@@ -310,7 +94,9 @@ class FilterChange(DeviceHvac):  # FAN: 10D0
             RQ, self.id, Code._10D0, PayloadT("00")
         )
 
-        if self._gwy.config.disable_discovery:  # discovery will poll for filter
+        if getattr(
+            self._gwy.config, "disable_discovery", False
+        ):  # discovery will poll for filter
             try:
                 asyncio.get_running_loop().call_soon(self.start_poller)
             except RuntimeError:
@@ -402,160 +188,14 @@ class RfsGateway(DeviceHvac):  # RFS: (spIDer gateway)
         if not hasattr(self, "hvac_state"):
             self.hvac_state = HvacState()
 
-        self.ctl = None
+        self.ctl: Device | None = None  # type: ignore[assignment]
         self._child_id = "hv"  # NOTE: domain_id
-        self.tcs = None
+        self.tcs: Evohome | None = None  # type: ignore[assignment]
 
     def _post_class_promote(self) -> None:
         """Initialize state when promoted from a generic HVAC device."""
         if not hasattr(self, "hvac_state"):
             self.hvac_state = HvacState()
-
-
-class HvacHumiditySensor(BatteryState, IndoorHumidity, Fakeable):  # HUM: I/12A0
-    """The class for a humidity sensor.
-
-    The cardinal code is 12A0.
-    """
-
-    _SLUG: str = DevType.HUM
-
-    async def temperature(self) -> float | None:
-        """Return the current temperature in Celsius.
-
-        :return: The temperature in degrees Celsius, or None if not available
-        :rtype: float | None
-        """
-        return self.hvac_state.temperature
-
-    async def dewpoint_temp(self) -> float | None:
-        """Return the dewpoint temperature in Celsius.
-
-        :return: The dewpoint temperature in degrees Celsius, or None if not available
-        :rtype: float | None
-        """
-        return self.hvac_state.dewpoint_temp
-
-    async def status(self) -> dict[str, Any]:
-        """Return the status of the humidity sensor.
-
-        :return: A dictionary containing the sensor's status including temperature and humidity
-        :rtype: dict[str, Any]
-        """
-        base_status = await super().status()
-        return {
-            **base_status,
-            SZ_TEMPERATURE: await self.temperature(),
-            "dewpoint_temp": await self.dewpoint_temp(),
-        }
-
-
-class HvacCarbonDioxideSensor(CarbonDioxide, Fakeable):  # CO2: I/1298
-    """The class for a CO2 sensor.
-
-    The cardinal code is 1298.
-    """
-
-    _SLUG: str = DevType.CO2
-
-    # .I --- 29:181813 63:262142 --:------ 1FC9 030 00-31E0-76C635 01-31E0-76C635 00-1298-76C635 67-10E0-76C635 00-1FC9-76C635
-    # .W --- 32:155617 29:181813 --:------ 1FC9 012 00-31D9-825FE1 00-31DA-825FE1  # The HRU
-    # .I --- 29:181813 32:155617 --:------ 1FC9 001 00
-
-    async def initiate_binding_process(self) -> Packet:
-        """Initiate the binding process for the CO2 sensor.
-
-        :return: The packet sent to initiate binding
-        :rtype: Packet
-        :raises exc.BindingError: If binding fails
-        """
-        return await super()._initiate_binding_process(
-            (Code._31E0, Code._1298, Code._2E10)
-        )
-
-
-class HvacRemote(BatteryState, Fakeable, HvacRemoteBase):  # REM: I/22F[138]
-    """The REM (remote/switch) class, such as a 4-way switch.
-
-    The cardinal codes are 22F1, 22F3 (also 22F8?).
-    """
-
-    _SLUG: str = DevType.REM
-
-    async def initiate_binding_process(self) -> Packet:
-        # .I --- 37:155617 --:------ 37:155617 1FC9 024 00-22F1-965FE1 00-22F3-965FE1 67-10E09-65FE1 00-1FC9-965FE1
-        # .W --- 32:155617 37:155617 --:------ 1FC9 012 00-31D9-825FE1 00-31DA-825FE1
-        # .I --- 37:155617 32:155617 --:------ 1FC9 001 00
-
-        return await super()._initiate_binding_process(
-            Code._22F1 if self._scheme == "nuaire" else (Code._22F1, Code._22F3)
-        )
-
-    async def fan_rate(self) -> str | None:
-        """Get the current fan rate setting.
-
-        :return: The fan rate as a string, or None if not available
-        :rtype: str | None
-        :note: This is a work in progress - rate can be either int or str
-        """
-        return self.hvac_state.fan_rate
-
-    async def set_fan_rate(self, value: int) -> Packet:
-        """Set a fake fan rate for the remote control.
-
-        :param value: The fan rate to set (can be int or str, but not None)
-        :type value: int
-        :raises TypeError: If the remote is not in faked mode
-        :return: The sent packet
-        :rtype: Packet
-        :note: This is a work in progress
-        """
-
-        if not self.is_faked:  # NOTE: some remotes are stateless (i.e. except seqn)
-            raise exc.DeviceNotFaked(f"{self}: Faking is not enabled")
-
-        # TODO: num_repeats=2, or wait_for_reply=True ?
-
-        # NOTE: this is not completely understood (i.e. diffs between vendor schemes)
-        cmd = Command.set_fan_mode(self.id, int(4 * value), src_id=self.id)
-        return await self._gwy.async_send_cmd(
-            cmd, num_repeats=2, priority=Priority.HIGH
-        )
-
-    async def fan_mode(self) -> str | None:
-        """Return the current fan mode.
-
-        :return: The fan mode as a string, or None if not available
-        :rtype: str | None
-        """
-        return self.hvac_state.fan_mode
-
-    async def boost_timer(self) -> int | None:
-        """Return the remaining boost timer in minutes.
-
-        :return: The remaining boost time in minutes, or None if boost is not active
-        :rtype: int | None
-        """
-        return self.hvac_state.boost_timer_mins
-
-    async def status(self) -> dict[str, Any]:
-        base_status = await super().status()
-        return {
-            **base_status,
-            SZ_FAN_MODE: await self.fan_mode(),
-            SZ_BOOST_TIMER: await self.boost_timer(),
-        }
-
-
-class HvacDisplayRemote(HvacRemote):  # DIS
-    """The DIS (display switch)."""
-
-    _SLUG: str = DevType.DIS
-
-    # async def initiate_binding_process(self) -> Packet:
-    #     return await super()._initiate_binding_process(
-    #         (Code._31E0, Code._1298, Code._2E10)
-    #     )
 
 
 class HvacVentilator(FilterChange):  # FAN: RP/31DA, I/31D[9A], 2411
@@ -967,7 +607,7 @@ class HvacVentilator(FilterChange):  # FAN: RP/31DA, I/31D[9A], 2411
         _LOGGER.debug("No bound REM or DIS devices found for FAN %s", self.id)
         return None
 
-    async def set_fan_mode(self, fan_mode: str | int) -> Packet:
+    async def set_fan_mode(self, fan_mode: str | int) -> Packet | None:
         """Set the operating mode/speed of the ventilator.
 
         :param fan_mode: The desired fan mode (e.g., 'low', 'medium', 'high', 'boost').
@@ -1288,190 +928,3 @@ class HvacVentilator(FilterChange):  # FAN: RP/31DA, I/31D[9A], 2411
 # class HvacFanPiv(HvacVentilator):
 #     """A positive input ventilation unit (aka: PIV)."""
 #     _SLUG: str = DEV_TYPE.PIV
-
-
-# e.g. {"HUM": HvacHumiditySensor}
-HVAC_CLASS_BY_SLUG: dict[str, type[DeviceHvac]] = class_by_attr(__name__, "_SLUG")
-
-
-def class_dev_hvac(
-    dev_addr: Address, *, msg: Message | None = None, eavesdrop: bool = False
-) -> type[DeviceHvac]:
-    """Return a device class, but only if the device must be from the HVAC group.
-
-    May return a base class, `DeviceHvac`, which will need promotion.
-    """
-
-    if not eavesdrop:
-        raise exc.DeviceNotRecognised(
-            f"No HVAC class for: {dev_addr} (no eavesdropping)"
-        )
-
-    if msg is None:
-        raise exc.DeviceNotRecognised(f"No HVAC class for: {dev_addr} (no msg)")
-
-    if klass := HVAC_KLASS_BY_VC_PAIR.get((msg.verb, msg.code)):
-        return HVAC_CLASS_BY_SLUG[klass]
-
-    if msg.code in CODES_OF_HVAC_DOMAIN_ONLY:
-        return DeviceHvac
-
-    raise exc.DeviceNotRecognised(
-        f"No HVAC class for: {dev_addr} (insufficient meta-data)"
-    )
-
-
-_REMOTES = {
-    "21800000": {
-        "name": "Orcon 15RF",
-        "mode": "1,2,3,T,Auto,Away",
-    },
-    "21800060": {
-        "name": "Orcon 15RF Display",
-        "mode": "1,2,3,T,Auto,Away",
-    },
-    "xxx": {
-        "name": "Orcon CO2 Control",
-        "mode": "1T,2T,3T,Auto,Away",
-    },
-    "03-00062": {
-        "name": "RFT-SPIDER",
-        "mode": "1,2,3,T,A",
-    },
-    "04-00045": {"name": "RFT-CO2"},  # mains-powered
-    "04-00046": {"name": "RFT-RV"},
-    "545-7550": {
-        "name": "RFT-PIR",
-    },
-    "536-0124": {  # idx="00"
-        "name": "RFT",
-        "mode": "1,2,3,T",
-        "CVE": False,  # not clear
-        "HRV": True,
-    },
-    "536-0146": {  # idx="??"
-        "name": "RFT-DF",
-        "mode": "",
-        "CVE": True,
-        "HRV": False,
-    },
-    "536-0150": {  # idx = "63"
-        "name": "RFT-AUTO",
-        "mode": "1,Auto,3,T",
-        "CVE": True,
-        "HRV": True,
-    },
-}
-
-
-# see: https://github.com/arjenhiemstra/ithowifi/blob/master/software/NRG_itho_wifi/src/IthoPacket.h
-
-"""
-Itho Remote (model) enums.
-
-CVE/HRU remote (536-0124) RFT W: 3 modes, timer
--------------------------------------------------
-
-.. table:: 536-0124
-   :widths: auto
-
-   ===========  =========================  ================================================
-    "away":     (Code._22F1, 00, 01|04"),  how to invoke?
-    "low":      (Code._22F1, 00, 02|04"),  aka eco
-    "medium":   (Code._22F1, 00, 03|04"),  aka auto (with sensors) - is that only for 63?
-    "high":     (Code._22F1, 00, 04|04"),  aka full
-
-    "timer_1":  (Code._22F3, 00, 00|0A"),  10 minutes full speed
-    "timer_2":  (Code._22F3, 00, 00|14"),  20 minutes full speed
-    "timer_3":  (Code._22F3, 00, 00|1E"),  30 minutes full speed
-   ===========  =========================  ================================================
-
-RFT-AUTO (536-0150) RFT CAR: 2 modes, auto, timer: idx = 63, essentially same as above, but also...
------------------------------------------------------------------------------------------------------
-
-.. table:: 536-0150
-   :widths: auto
-
-   =============  =========================  ================================================
-   "auto_night":  (Code._22F8, 63, 02|03"),  additional - press auto x2
-   =============  =========================  ================================================
-
-RFT-RV (04-00046), RFT-CO2 (04-00045) - sensors with control
-------------------------------------------------------------
-
-.. table:: 04-00046
-   :widths: auto
-
-   ==============  ========================================   =============
-    "medium":      (Code._22F1, 00, 03|07"),                  1=away, 2=low?
-    "auto":        (Code._22F1, 00, 05|07"),                  4=high
-    "auto_night":  (Code._22F1, 00, 0B|0B"),
-
-    "timer_1":     (Code._22F3, 00, 00|0A, 00|00, 0000"),     10 minutes
-    "timer_2":     (Code._22F3, 00, 00|14, 00|00, 0000"),     20 minutes
-    "timer_3":     (Code._22F3, 00, 00|1E, 00|00, 0000"),     30 minutes
-   ==============  ========================================   =============
-
-RFT-PIR (545-7550) - presence sensor
-------------------------------------
-
-RFT_DF: DemandFlow remote (536-0146)
-------------------------------------
-
-.. table:: 536-0146
-   :widths: auto
-
-   ===========  ================================  =========================================
-    "timer_1":  (Code._22F3, 00, 42|03, 03|03"),  0b01-000-010 = 3 hrs, back to last mode
-    "timer_2":  (Code._22F3, 00, 42|06, 03|03"),  0b01-000-010 = 6 hrs, back to last mode
-    "timer_3":  (Code._22F3, 00, 42|09, 03|03"),  0b01-000-010 = 9 hrs, back to last mode
-    "cook_30":  (Code._22F3, 00, 02|1E, 02|03"),  30 mins (press 1x)
-    "cook_60":  (Code._22F3, 00, 02|3C, 02|03"),  60 mins (press 2x)
-
-    "low":      (Code._22F8, 00, 01|02"),         ?eco     co2 <= 1200 ppm?
-    "high":     (Code._22F8, 00, 02|02"),         ?comfort co2 <= 1000 ppm?
-   ===========  ================================  =========================================
-
-
-Join commands:
---------------
-
-.. table:: join per accessory type
-   :widths: auto
-
-   ==========  =================  =====================  =========================  ==========================  =========================  ==========================  =================  ==========
-   type        set 1              set 2                  set 3                      set 4                       set 5                      set 6                       description        art #
-   ==========  =================  =====================  =========================  ==========================  =========================  ==========================  =================  ==========
-   "CVERFT":   (Code._1FC9,  00,  Code._22F1, 0x000000,                             01, Code._10E0, 0x000000")                                                         CVE/HRU remote     (536-0124)
-   "AUTORFT":  (Code._1FC9,  63,  Code._22F8, 0x000000,                             01, Code._10E0, 0x000000")                                                         AUTO RFT           (536-0150)
-   "DF":       (Code._1FC9,  00,  Code._22F8, 0x000000,                             00, Code._10E0, 0x000000")                                                         DemandFlow remote  (536-0146)
-   "RV":       (Code._1FC9,  00,  Code._12A0, 0x000000,                             01, Code._10E0, 0x000000,   00, Code._31E0, 0x000000,  00, Code._1FC9, 0x000000")  RFT-RV             (04-00046)
-   "CO2":      (Code._1FC9,  00,  Code._1298, 0x000000,  00, Code._2E10, 0x000000,  01, Code._10E0, 0x000000,   00, Code._31E0, 0x000000,  00, Code._1FC9, 0x000000")  RFT-CO2            (04-00045)
-   ==========  =================  =====================  =========================  ==========================  =========================  ==========================  =================  ==========
-
-Leave commands:
----------------
-
-.. table:: leave per accessory type
-   :widths: auto
-
-   ==========  =================  ======================  =========================  ==========
-   type        set 1              set 2                   description                art #
-   ==========  =================  ======================  =========================  ==========
-   "Others":   (Code._1FC9, 00,   Code._1FC9, 0x000000")  standard leave command
-   "AUTORFT":  (Code._1FC9, 63,   Code._1FC9, 0x000000")  leave command of AUTO RFT  (536-0150)
-   ==========  =================  ======================  =========================  ==========
-
-.. table:: verbs
-   :widths: 2, 4
-
-   ======  ========
-   verb    byte
-   ======  ========
-   ``RQ``  ``0x00``
-   ``I_``  ``0x01``
-   ``W_``  ``0x02``
-   ``RP``  ``0x03``
-   ======  ========
-
-"""
