+++ src/ramses_rf/devices/opentherm_bridge.py
@@ -1,0 +1,733 @@
+"""RAMSES RF - OpenTherm Bridge Device."""
+
+from __future__ import annotations
+
+import logging
+from collections.abc import Callable
+from datetime import timedelta as td
+from typing import TYPE_CHECKING, Any, Final, Literal, cast
+
+from ramses_rf.const import (
+    FC,
+    HEARTBEAT_TIMEOUT_OTB,
+    I_,
+    RP,
+    RQ,
+    SZ_HEAT_DEMAND,
+    SZ_PRESSURE,
+    SZ_SETPOINT,
+    SZ_TEMPERATURE,
+    Code,
+    DevType,
+)
+from ramses_rf.models import DemandState, DeviceTraits, OpenThermState, TemperatureState
+from ramses_rf.quirks import QUARANTINED_OT_MSG_IDS
+from ramses_tx import Command, Priority
+from ramses_tx.const import (
+    SZ_BOILER_OUTPUT_TEMP,
+    SZ_BOILER_RETURN_TEMP,
+    SZ_BOILER_SETPOINT,
+    SZ_BURNER_FAILED_STARTS,
+    SZ_BURNER_HOURS,
+    SZ_BURNER_STARTS,
+    SZ_CH_ACTIVE,
+    SZ_CH_ENABLED,
+    SZ_CH_MAX_SETPOINT,
+    SZ_CH_PUMP_HOURS,
+    SZ_CH_PUMP_STARTS,
+    SZ_CH_SETPOINT,
+    SZ_CH_WATER_PRESSURE,
+    SZ_COOLING_ACTIVE,
+    SZ_COOLING_ENABLED,
+    SZ_DHW_ACTIVE,
+    SZ_DHW_BLOCKING,
+    SZ_DHW_BURNER_HOURS,
+    SZ_DHW_BURNER_STARTS,
+    SZ_DHW_ENABLED,
+    SZ_DHW_FLOW_RATE,
+    SZ_DHW_PUMP_HOURS,
+    SZ_DHW_PUMP_STARTS,
+    SZ_DHW_SETPOINT,
+    SZ_DHW_TEMP,
+    SZ_FAULT_PRESENT,
+    SZ_FLAME_ACTIVE,
+    SZ_FLAME_SIGNAL_LOW,
+    SZ_MAX_REL_MODULATION,
+    SZ_NUM_REPEATS,
+    SZ_OEM_CODE,
+    SZ_OTC_ACTIVE,
+    SZ_OUTSIDE_TEMP,
+    SZ_PRIORITY,
+    SZ_REL_MODULATION_LEVEL,
+    SZ_SUMMER_MODE,
+    MsgId,
+)
+from ramses_tx.typing import PayDictT, PayloadT
+
+from ..protocol.opentherm import (
+    PARAMS_DATA_IDS,
+    SCHEMA_DATA_IDS,
+    STATUS_DATA_IDS,
+    SZ_MSG_ID,
+    SZ_MSG_NAME,
+    SZ_MSG_TYPE,
+    SZ_VALUE,
+    OtDataId,
+    OtMsgType,
+)
+from .heat_actuators import Actuator, HeatDemand
+
+if TYPE_CHECKING:
+    from ..messages import Message
+
+QOS_LOW = {SZ_PRIORITY: Priority.LOW}  # FIXME:  deprecate QoS in kwargs
+QOS_MID = {SZ_PRIORITY: Priority.HIGH}  # FIXME: deprecate QoS in kwargs
+QOS_MAX = {SZ_PRIORITY: Priority.HIGH, SZ_NUM_REPEATS: 3}  # FIXME: deprecate QoS...
+
+#
+# NOTE: All debug flags should be False for deployment to end-users
+_DBG_ENABLE_DEPRECATION: Final[bool] = False
+_DBG_EXTRA_OTB_DISCOVERY: Final[bool] = False
+
+_LOGGER = logging.getLogger(__name__)
+
+
+def _to_msg_id(data_id: OtDataId) -> MsgId:
+    return cast(MsgId, f"{data_id:02X}")
+
+
+# NOTE: config.use_native_ot should enforce sends, but not reads from _msgz DB
+class OtbGateway(Actuator, HeatDemand):  # OTB (10): 3220 (22D9, others)
+    """The OTB class, specifically an OpenTherm Bridge (R8810A Bridge)."""
+
+    # see: https://www.opentherm.eu/request-details/?post_ids=2944
+    # see: https://www.automatedhome.co.uk/vbulletin/showthread.php?6400-(New)-cool-mode-in-Evohome
+
+    _SLUG = DevType.OTB
+    _STATE_ATTR = SZ_REL_MODULATION_LEVEL
+
+    OT_TO_RAMSES: dict[MsgId, Code] = {  # TODO: move to opentherm.py
+        MsgId._00: Code._3EF0,  # master/slave status (actuator_state)
+        MsgId._01: Code._22D9,  # boiler_setpoint
+        MsgId._0E: Code._3EF0,  # max_rel_modulation_level (is a PARAM?)
+        MsgId._11: Code._3EF0,  # rel_modulation_level (actuator_state, also Code._3EF1)
+        MsgId._12: Code._1300,  # ch_water_pressure
+        MsgId._13: Code._12F0,  # dhw_flow_rate
+        MsgId._19: Code._3200,  # boiler_output_temp
+        MsgId._1A: Code._1260,  # dhw_temp
+        MsgId._1B: Code._1290,  # outside_temp
+        MsgId._1C: Code._3210,  # boiler_return_temp
+        MsgId._38: Code._10A0,  # dhw_setpoint (is a PARAM)
+        MsgId._39: Code._1081,  # ch_max_setpoint (is a PARAM)
+    }
+    RAMSES_TO_OT: dict[Code, MsgId] = {
+        v: k for k, v in OT_TO_RAMSES.items() if v != Code._3EF0
+    }  # also 10A0?
+
+    def __init__(
+        self, *args: Any, traits: DeviceTraits | None = None, **kwargs: Any
+    ) -> None:
+        """Initialize the OpenTherm Bridge device software twin.
+
+        :param args: Positional arguments passed to base class.
+        :param traits: Strictly typed traits definition object.
+        :type traits: DeviceTraits | None
+        :param kwargs: Keyword arguments passed to base class.
+        """
+        super().__init__(*args, traits=traits, **kwargs)
+        self.opentherm_state = OpenThermState()
+
+        self._child_id = FC  # NOTE: domain_id
+        self._msgs_ot: dict[MsgId, Message] = {}
+
+    def _post_class_promote(self) -> None:
+        """Initialize OTB state when promoted in-place from a generic device."""
+        self.__dict__.setdefault("_child_id", FC)
+        self.__dict__.setdefault("_msgs_ot", {})
+
+        if not hasattr(self, "temp_state"):
+            self.temp_state = TemperatureState()
+        if not hasattr(self, "demand_state"):
+            self.demand_state = DemandState()
+        if not hasattr(self, "opentherm_state"):
+            self.opentherm_state = OpenThermState()
+
+    @property
+    def heartbeat_timeout(self) -> td:
+        """Return the timeout before the device is considered unavailable.
+
+        :return: The timeout duration.
+        :rtype: td
+        """
+        return HEARTBEAT_TIMEOUT_OTB
+
+    def _setup_discovery_cmds(self) -> None:
+        def which_cmd(
+            use_native_ot: Literal["always", "prefer", "avoid", "never"] | str | None,
+            msg_id: MsgId,
+        ) -> Command | None:
+            """Create a OT cmd, or its RAMSES equivalent, depending."""
+            # we know RQ|3220 is an option, question is: use that, or RAMSES or nothing?
+            if use_native_ot in ("always", "prefer"):
+                return Command.get_opentherm_data(self.id, msg_id)
+            if msg_id in self.OT_TO_RAMSES:  # is: in ("avoid", "never")
+                return Command.from_attrs(
+                    RQ, self.id, self.OT_TO_RAMSES[msg_id], PayloadT("00")
+                )
+            if use_native_ot == "avoid":
+                return Command.get_opentherm_data(self.id, msg_id)
+            return None  # use_native_ot == "never"
+
+        super()._setup_discovery_cmds()
+
+        # always send at least one of RQ|3EF0 or RQ|3220|00 (status)
+        if getattr(self._gwy.config, "use_native_ot", "avoid") != "never":
+            self.discovery.add_cmd(Command.get_opentherm_data(self.id, MsgId._00), 60)
+
+        if getattr(self._gwy.config, "use_native_ot", "avoid") != "always":
+            self.discovery.add_cmd(
+                Command.from_attrs(RQ, self.id, Code._3EF0, PayloadT("00")), 60
+            )
+            self.discovery.add_cmd(  # NOTE: this code is a WIP
+                Command.from_attrs(RQ, self.id, Code._2401, PayloadT("00")), 60
+            )
+
+        for data_id in SCHEMA_DATA_IDS:  # From OT v2.2: version numbers
+            if cmd := which_cmd(
+                getattr(self._gwy.config, "use_native_ot", "avoid"), _to_msg_id(data_id)
+            ):
+                self.discovery.add_cmd(cmd, 6 * 3600, delay=180)
+
+        for data_id in PARAMS_DATA_IDS:  # params or L/T state
+            if cmd := which_cmd(
+                getattr(self._gwy.config, "use_native_ot", "avoid"), _to_msg_id(data_id)
+            ):
+                self.discovery.add_cmd(cmd, 3600, delay=90)
+
+        for data_id in STATUS_DATA_IDS:  # except "00", see above
+            if data_id == 0x00:
+                continue
+            if cmd := which_cmd(
+                getattr(self._gwy.config, "use_native_ot", "avoid"), _to_msg_id(data_id)
+            ):
+                self.discovery.add_cmd(cmd, 300, delay=15)
+
+        if _DBG_EXTRA_OTB_DISCOVERY:  # TODO: these are WIP, but do vary in payload
+            for code in (
+                Code._2401,  # WIP - modulation_level + flags?
+                Code._3221,  # R8810A/20A
+                Code._3223,  # R8810A/20A
+            ):
+                self.discovery.add_cmd(
+                    Command.from_attrs(RQ, self.id, code, PayloadT("00")), 60
+                )
+
+        if _DBG_EXTRA_OTB_DISCOVERY:  # TODO: these are WIP, appear FIXED in payload
+            for code in (
+                Code._0150,  # payload always "000000", R8820A only?
+                Code._1098,  # payload always "00C8",   R8820A only?
+                Code._10B0,  # payload always "0000",   R8820A only?
+                Code._1FD0,  # payload always "0000000000000000"
+                Code._2400,  # payload always "0000000F"
+                Code._2410,  # payload always "000000000000000000000000010000000100000C"
+                Code._2420,  # payload always "0000001000000...
+            ):  # TODO: to test against BDR91T
+                self.discovery.add_cmd(
+                    Command.from_attrs(RQ, self.id, code, PayloadT("00")), 300
+                )
+
+    def _handle_msg(self, msg: Message) -> None:
+        super()._handle_msg(msg)
+
+        if msg.verb not in (I_, RP):
+            return
+
+        if msg.code == Code._3220:
+            self._handle_3220(msg)
+        elif msg.code in self.RAMSES_TO_OT:
+            self._handle_code(msg)
+
+    def _handle_3220(self, msg: Message) -> None:
+        """Handle 3220-based messages."""
+
+        # NOTE: Reserved msgs have null data, but that msg_id may later be OK!
+        if msg.payload[SZ_MSG_TYPE] == OtMsgType.RESERVED:
+            return
+
+        # NOTE: Some msgs have invalid data, but that msg_id may later be OK!
+        if msg.payload.get(SZ_VALUE) is None:
+            return
+
+        # msg_id is int in msg payload/opentherm.py, but MsgId (str) in this module
+        msg_id = _to_msg_id(msg.payload[SZ_MSG_ID])
+        self._msgs_ot[msg_id] = msg
+
+        if not _DBG_ENABLE_DEPRECATION:  # FIXME: data gaps
+            return
+
+        reset = msg.payload[SZ_MSG_TYPE] not in (
+            OtMsgType.DATA_INVALID,
+            OtMsgType.UNKNOWN_DATAID,
+            OtMsgType.RESERVED,  # but some are ?always reserved
+        )
+        self.discovery.deprecate_code_ctx(msg._pkt, ctx=msg_id, reset=reset)
+
+    def _handle_code(self, msg: Message) -> None:
+        """Handle non-3220-based messages."""
+
+        if msg.code == Code._3EF0 and msg.verb == I_:
+            # NOTE: this is development/discovery code  # chasing flags
+            # self._send_cmd(
+            #     Command.get_opentherm_data(self.id, MsgId._00), **QOS_MID
+            # )  # FIXME: deprecate QoS in kwargs
+            return
+
+        if msg.code in (Code._10A0, Code._3EF1):
+            return
+
+        if not _DBG_ENABLE_DEPRECATION:  # FIXME: data gaps
+            return
+
+        # TODO: can be temporarily 7FFF?
+        if msg._pkt.payload[2:] == "7FFF" or (
+            msg.code == Code._1300 and msg._pkt.payload[2:] == "09F6"
+        ):  # latter is CH water pressure
+            self.discovery.deprecate_code_ctx(msg._pkt)
+        else:
+            self.discovery.deprecate_code_ctx(msg._pkt, reset=True)
+
+    def _ot_msg_flag(self, msg_id: MsgId, flag_idx: int) -> bool | None:
+        flags = cast(list[int], self._ot_msg_value(msg_id))
+        return bool(flags[flag_idx]) if flags else None
+
+    @staticmethod
+    def _ot_msg_name(msg: Message) -> str:  # TODO: remove
+        return (
+            msg.payload[SZ_MSG_NAME]
+            if isinstance(msg.payload[SZ_MSG_NAME], str)
+            else f"{msg.payload[SZ_MSG_ID]:02X}"
+        )
+
+    def _ot_msg_value(self, msg_id: MsgId) -> int | float | list[int] | None:
+        if msg_id in QUARANTINED_OT_MSG_IDS.get(self._SLUG, set()):
+            return None
+        # data_id = int(msg_id, 16)
+        if (msg := self._msgs_ot.get(msg_id)) and not getattr(msg, "_expired", False):
+            # TODO: value_hb/_lb
+            return msg.payload.get(SZ_VALUE)  # type: ignore[no-any-return]
+        return None
+
+    def _result_by_callback(
+        self, cbk_ot: Callable[[], Any] | None, cbk_ramses: Callable[[], Any] | None
+    ) -> Any | None:
+        """Return a value using OpenTherm or RAMSES as per `config.use_native_ot`."""
+        use_ot = getattr(self._gwy.config, "use_native_ot", "avoid")
+        if use_ot == "always":
+            return cbk_ot() if cbk_ot else None
+        if use_ot == "prefer":
+            if cbk_ot and (result := cbk_ot()) is not None:
+                return result
+
+        result_ramses = cbk_ramses() if cbk_ramses is not None else None
+        if use_ot == "avoid" and result_ramses is None:
+            return cbk_ot() if cbk_ot else None
+        return result_ramses  # incl. use_native_ot == "never"
+
+    async def _result_by_lookup(
+        self,
+        code: Code,
+        /,
+        *,
+        key: str,
+    ) -> Any | None:
+        """Return a value using OpenTherm or RAMSES as per `config.use_native_ot`."""
+        use_ot = getattr(self._gwy.config, "use_native_ot", "avoid")
+
+        if use_ot in ("always", "prefer"):
+            if (result_ot := self._ot_msg_value(self.RAMSES_TO_OT[code])) is not None:
+                return result_ot
+
+        result_ramses = await self.entity_state.get_value(code, key=key)
+        if result_ramses is None and use_ot != "never":
+            return self._ot_msg_value(self.RAMSES_TO_OT[code])
+
+        return result_ramses  # incl. use_native_ot == "never"
+
+    def _result_by_value(
+        self, result_ot: Any | None, result_ramses: Any | None
+    ) -> Any | None:
+        """Return a value using OpenTherm or RAMSES as per `config.use_native_ot`."""
+        use_ot = getattr(self._gwy.config, "use_native_ot", "avoid")
+
+        if use_ot in ("always", "prefer"):
+            if result_ot is not None:
+                return result_ot
+
+        if result_ramses is None and use_ot != "never":
+            return result_ot
+
+        return result_ramses  # incl. use_native_ot == "never"
+
+    async def bit_2_4(self) -> bool | None:  # 2401 - WIP
+        return await self.entity_state.get_flag(Code._2401, "_flags_2", 4)
+
+    async def bit_2_5(self) -> bool | None:  # 2401 - WIP
+        return await self.entity_state.get_flag(Code._2401, "_flags_2", 5)
+
+    async def bit_2_6(self) -> bool | None:  # 2401 - WIP
+        return await self.entity_state.get_flag(Code._2401, "_flags_2", 6)
+
+    async def bit_2_7(self) -> bool | None:  # 2401 - WIP
+        return await self.entity_state.get_flag(Code._2401, "_flags_2", 7)
+
+    async def bit_3_7(self) -> bool | None:  # 3EF0 (byte 3, only OTB)
+        return await self.entity_state.get_flag(Code._3EF0, "_flags_3", 7)
+
+    async def bit_6_6(self) -> bool | None:  # 3EF0 ?dhw_enabled (byte 3, only R8820A?)
+        return await self.entity_state.get_flag(Code._3EF0, "_flags_6", 6)
+
+    async def percent(self) -> float | None:  # 2401 - WIP (~3150|FC)
+        return cast(
+            float | None,
+            await self.entity_state.get_value(Code._2401, key=SZ_HEAT_DEMAND),
+        )
+
+    async def value(self) -> int | None:  # 2401 - WIP
+        return cast(
+            int | None, await self.entity_state.get_value(Code._2401, key="_value_2")
+        )
+
+    async def boiler_output_temp(self) -> float | None:  # 3220|19, or 3200
+        return cast(
+            float | None, await self._result_by_lookup(Code._3200, key=SZ_TEMPERATURE)
+        )
+
+    async def boiler_return_temp(self) -> float | None:  # 3220|1C, or 3210
+        return cast(
+            float | None, await self._result_by_lookup(Code._3210, key=SZ_TEMPERATURE)
+        )
+
+    async def boiler_setpoint(self) -> float | None:  # 3220|01, or 22D9
+        return cast(
+            float | None, await self._result_by_lookup(Code._22D9, key=SZ_SETPOINT)
+        )
+
+    async def ch_max_setpoint(self) -> float | None:  # 3220|39, or 1081
+        return cast(
+            float | None, await self._result_by_lookup(Code._1081, key=SZ_SETPOINT)
+        )
+
+    async def ch_setpoint(self) -> float | None:  # 3EF0 (byte 7, only R8820A?)
+        return cast(
+            float | None,
+            self._result_by_value(
+                None, await self.entity_state.get_value(Code._3EF0, key=SZ_CH_SETPOINT)
+            ),
+        )
+
+    async def ch_water_pressure(self) -> float | None:  # 3220|12, or 1300
+        return cast(
+            float | None, await self._result_by_lookup(Code._1300, key=SZ_PRESSURE)
+        )
+
+    async def dhw_flow_rate(self) -> float | None:  # 3220|13, or 12F0
+        return cast(
+            float | None, await self._result_by_lookup(Code._12F0, key=SZ_DHW_FLOW_RATE)
+        )
+
+    async def dhw_setpoint(self) -> float | None:  # 3220|38, or 10A0
+        return cast(
+            float | None, await self._result_by_lookup(Code._10A0, key=SZ_SETPOINT)
+        )
+
+    async def dhw_temp(self) -> float | None:  # 3220|1A, or 1260
+        return cast(
+            float | None, await self._result_by_lookup(Code._1260, key=SZ_TEMPERATURE)
+        )
+
+    async def max_rel_modulation(self) -> float | None:  # 3220|0E, or 3EF0 (byte 8)
+        return cast(
+            float | None,
+            self._result_by_value(
+                self._ot_msg_value(MsgId._0E),  # NOTE: not reliable?
+                await self.entity_state.get_value(
+                    Code._3EF0, key=SZ_MAX_REL_MODULATION
+                ),
+            ),
+        )
+
+    async def oem_code(self) -> float | None:  # 3220|73, no known RAMSES equivalent
+        return cast(float | None, self._ot_msg_value(MsgId._73))
+
+    async def outside_temp(self) -> float | None:  # 3220|1B, 1290
+        return cast(
+            float | None, await self._result_by_lookup(Code._1290, key=SZ_TEMPERATURE)
+        )
+
+    async def rel_modulation_level(self) -> float | None:  # 3220|11, or 3EF0/3EF1
+        return cast(
+            float | None,
+            self._result_by_value(
+                self._ot_msg_value(MsgId._11),  # NOTE: not reliable?
+                await self.entity_state.get_value(
+                    (Code._3EF0, Code._3EF1), key=self.MODULATION_LEVEL
+                ),
+            ),
+        )
+
+    async def ch_active(self) -> bool | None:  # 3220|00, or 3EF0 (byte 3)
+        return cast(
+            bool | None,
+            self._result_by_value(
+                self._ot_msg_flag(MsgId._00, 8 + 1),  # NOTE: not reliable?
+                await self.entity_state.get_value(Code._3EF0, key=SZ_CH_ACTIVE),
+            ),
+        )
+
+    async def ch_enabled(self) -> bool | None:  # 3220|00, or 3EF0 (byte 6)
+        return cast(
+            bool | None,
+            self._result_by_value(
+                self._ot_msg_flag(MsgId._00, 0),  # NOTE: not reliable?
+                await self.entity_state.get_value(Code._3EF0, key=SZ_CH_ENABLED),
+            ),
+        )
+
+    async def cooling_active(self) -> bool | None:  # 3220|00, TODO: no known RAMSES
+        return cast(
+            bool | None,
+            self._result_by_value(self._ot_msg_flag(MsgId._00, 8 + 4), None),
+        )
+
+    async def cooling_enabled(self) -> bool | None:  # 3220|00, TODO: no known RAMSES
+        return cast(
+            bool | None, self._result_by_value(self._ot_msg_flag(MsgId._00, 2), None)
+        )
+
+    async def dhw_active(self) -> bool | None:  # 3220|00, or 3EF0 (byte 3)
+        return cast(
+            bool | None,
+            self._result_by_value(
+                self._ot_msg_flag(MsgId._00, 8 + 2),  # NOTE: not reliable?
+                await self.entity_state.get_value(Code._3EF0, key=SZ_DHW_ACTIVE),
+            ),
+        )
+
+    async def dhw_blocking(self) -> bool | None:  # 3220|00, TODO: no known RAMSES
+        return cast(
+            bool | None, self._result_by_value(self._ot_msg_flag(MsgId._00, 6), None)
+        )
+
+    async def dhw_enabled(self) -> bool | None:  # 3220|00, TODO: no known RAMSES
+        return cast(
+            bool | None, self._result_by_value(self._ot_msg_flag(MsgId._00, 1), None)
+        )
+
+    async def fault_present(self) -> bool | None:  # 3220|00, TODO: no known RAMSES
+        return cast(
+            bool | None, self._result_by_value(self._ot_msg_flag(MsgId._00, 8), None)
+        )
+
+    async def flame_active(self) -> bool | None:  # 3220|00, or 3EF0 (byte 3)
+        return cast(
+            bool | None,
+            self._result_by_value(
+                self._ot_msg_flag(MsgId._00, 8 + 3),  # NOTE: not reliable?
+                await self.entity_state.get_value(Code._3EF0, key="flame_on"),
+            ),
+        )
+
+    async def otc_active(self) -> bool | None:  # 3220|00, TODO: no known RAMSES
+        return cast(
+            bool | None, self._result_by_value(self._ot_msg_flag(MsgId._00, 3), None)
+        )
+
+    async def summer_mode(self) -> bool | None:  # 3220|00, TODO: no known RAMSES
+        return cast(
+            bool | None, self._result_by_value(self._ot_msg_flag(MsgId._00, 5), None)
+        )
+
+    async def opentherm_schema(self) -> dict[str, Any]:
+        result: dict[str, Any] = {
+            self._ot_msg_name(v): v.payload
+            for k, v in self._msgs_ot.items()
+            if getattr(self.discovery, "_supported_cmds_ctx", {}).get(k)
+            and int(k, 16) in SCHEMA_DATA_IDS
+        }
+        return {
+            m: {k: v for k, v in p.items() if k.startswith(SZ_VALUE)}
+            for m, p in result.items()
+        }
+
+    async def opentherm_counters(self) -> dict[str, Any]:  # all are U16
+        return {
+            SZ_BURNER_HOURS: self._ot_msg_value(MsgId._78),
+            SZ_BURNER_STARTS: self._ot_msg_value(MsgId._74),
+            SZ_BURNER_FAILED_STARTS: self._ot_msg_value(MsgId._71),
+            SZ_CH_PUMP_HOURS: self._ot_msg_value(MsgId._79),
+            SZ_CH_PUMP_STARTS: self._ot_msg_value(MsgId._75),
+            SZ_DHW_BURNER_HOURS: self._ot_msg_value(MsgId._7B),
+            SZ_DHW_BURNER_STARTS: self._ot_msg_value(MsgId._77),
+            SZ_DHW_PUMP_HOURS: self._ot_msg_value(MsgId._7A),
+            SZ_DHW_PUMP_STARTS: self._ot_msg_value(MsgId._76),
+            SZ_FLAME_SIGNAL_LOW: self._ot_msg_value(MsgId._72),
+        }  # 0x73 is not a counter: is OEM diagnostic code...
+
+    async def opentherm_params(
+        self,
+    ) -> dict[str, Any]:  # F8_8, U8, {"hb": S8, "lb": S8}
+        result = {
+            self._ot_msg_name(v): v.payload
+            for k, v in self._msgs_ot.items()
+            if getattr(self.discovery, "_supported_cmds_ctx", {}).get(k)
+            and int(k, 16) in PARAMS_DATA_IDS
+        }
+        return {
+            m: {k: v for k, v in p.items() if k.startswith(SZ_VALUE)}
+            for m, p in result.items()
+        }
+
+    async def opentherm_status(
+        self,
+    ) -> dict[str, Any]:  # F8_8, U16 (only OEM_CODE) or bool
+        return {  # most these are in: STATUS_DATA_IDS
+            SZ_BOILER_OUTPUT_TEMP: self._ot_msg_value(MsgId._19),
+            SZ_BOILER_RETURN_TEMP: self._ot_msg_value(MsgId._1C),
+            SZ_BOILER_SETPOINT: self._ot_msg_value(MsgId._01),
+            # SZ_CH_MAX_SETPOINT: self._ot_msg_value(MsgId._39),  # in PARAMS_DATA_IDS
+            SZ_CH_WATER_PRESSURE: self._ot_msg_value(MsgId._12),
+            SZ_DHW_FLOW_RATE: self._ot_msg_value(MsgId._13),
+            # SZ_DHW_SETPOINT: self._ot_msg_value(MsgId._38),  # in PARAMS_DATA_IDS
+            SZ_DHW_TEMP: self._ot_msg_value(MsgId._1A),
+            SZ_OEM_CODE: self._ot_msg_value(MsgId._73),
+            SZ_OUTSIDE_TEMP: self._ot_msg_value(MsgId._1B),
+            SZ_REL_MODULATION_LEVEL: self._ot_msg_value(MsgId._11),
+            #
+            # SZ...: self._ot_msg_value(MsgId._05),  # in STATUS_DATA_IDS
+            # SZ...: self._ot_msg_value(MsgId._18),  # in STATUS_DATA_IDS
+            #
+            SZ_CH_ACTIVE: self._ot_msg_flag(MsgId._00, 8 + 1),
+            SZ_CH_ENABLED: self._ot_msg_flag(MsgId._00, 0),
+            SZ_COOLING_ACTIVE: self._ot_msg_flag(MsgId._00, 8 + 4),
+            SZ_COOLING_ENABLED: self._ot_msg_flag(MsgId._00, 2),
+            SZ_DHW_ACTIVE: self._ot_msg_flag(MsgId._00, 8 + 2),
+            SZ_DHW_BLOCKING: self._ot_msg_flag(MsgId._00, 6),
+            SZ_DHW_ENABLED: self._ot_msg_flag(MsgId._00, 1),
+            SZ_FAULT_PRESENT: self._ot_msg_flag(MsgId._00, 8),
+            SZ_FLAME_ACTIVE: self._ot_msg_flag(MsgId._00, 8 + 3),
+            SZ_SUMMER_MODE: self._ot_msg_flag(MsgId._00, 5),
+            SZ_OTC_ACTIVE: self._ot_msg_flag(MsgId._00, 3),
+        }
+
+    async def ramses_schema(self) -> PayDictT.EMPTY:
+        return {}
+
+    async def ramses_params(self) -> dict[str, float | None]:
+        return {
+            SZ_MAX_REL_MODULATION: await self.max_rel_modulation(),
+        }
+
+    async def ramses_status(self) -> dict[str, Any]:
+        return {
+            SZ_BOILER_OUTPUT_TEMP: await self.entity_state.get_value(
+                Code._3200, key=SZ_TEMPERATURE
+            ),
+            SZ_BOILER_RETURN_TEMP: await self.entity_state.get_value(
+                Code._3210, key=SZ_TEMPERATURE
+            ),
+            SZ_BOILER_SETPOINT: await self.entity_state.get_value(
+                Code._22D9, key=SZ_SETPOINT
+            ),
+            SZ_CH_MAX_SETPOINT: await self.entity_state.get_value(
+                Code._1081, key=SZ_SETPOINT
+            ),
+            SZ_CH_SETPOINT: await self.entity_state.get_value(
+                Code._3EF0, key=SZ_CH_SETPOINT
+            ),
+            SZ_CH_WATER_PRESSURE: await self.entity_state.get_value(
+                Code._1300, key=SZ_PRESSURE
+            ),
+            SZ_DHW_FLOW_RATE: await self.entity_state.get_value(
+                Code._12F0, key=SZ_DHW_FLOW_RATE
+            ),
+            SZ_DHW_SETPOINT: await self.entity_state.get_value(
+                Code._1300, key=SZ_SETPOINT
+            ),
+            SZ_DHW_TEMP: await self.entity_state.get_value(
+                Code._1260, key=SZ_TEMPERATURE
+            ),
+            SZ_OUTSIDE_TEMP: await self.entity_state.get_value(
+                Code._1290, key=SZ_TEMPERATURE
+            ),
+            SZ_REL_MODULATION_LEVEL: await self.entity_state.get_value(
+                (Code._3EF0, Code._3EF1), key=self.MODULATION_LEVEL
+            ),
+            SZ_CH_ACTIVE: await self.entity_state.get_value(
+                Code._3EF0, key=SZ_CH_ACTIVE
+            ),
+            SZ_CH_ENABLED: await self.entity_state.get_value(
+                Code._3EF0, key=SZ_CH_ENABLED
+            ),
+            SZ_DHW_ACTIVE: await self.entity_state.get_value(
+                Code._3EF0, key=SZ_DHW_ACTIVE
+            ),
+            SZ_FLAME_ACTIVE: await self.entity_state.get_value(
+                Code._3EF0, key=SZ_FLAME_ACTIVE
+            ),
+        }
+
+    async def traits(self) -> dict[str, Any]:
+        base_traits = await super().traits()
+        return {
+            **base_traits,
+            "opentherm_traits": await self.discovery.supported_cmds_ot(),
+            "ramses_ii_traits": await self.discovery.supported_cmds(),
+        }
+
+    async def schema(self) -> dict[str, Any]:
+        base_schema = await super().schema()
+        return {
+            **base_schema,
+            "opentherm_schema": await self.opentherm_schema(),
+            "ramses_ii_schema": await self.ramses_schema(),
+        }
+
+    async def params(self) -> dict[str, Any]:
+        base_params = await super().params()
+        return {
+            **base_params,
+            "opentherm_params": await self.opentherm_params(),
+            "ramses_ii_params": await self.ramses_params(),
+        }
+
+    async def status(self) -> dict[str, Any]:
+        base_status = await super().status()
+        return {
+            **base_status,  # incl. actuator_cycle, actuator_state
+            SZ_BOILER_OUTPUT_TEMP: await self.boiler_output_temp(),
+            SZ_BOILER_RETURN_TEMP: await self.boiler_return_temp(),
+            SZ_BOILER_SETPOINT: await self.boiler_setpoint(),
+            SZ_CH_SETPOINT: await self.ch_setpoint(),
+            SZ_CH_MAX_SETPOINT: await self.ch_max_setpoint(),
+            SZ_CH_WATER_PRESSURE: await self.ch_water_pressure(),
+            SZ_DHW_FLOW_RATE: await self.dhw_flow_rate(),
+            SZ_DHW_SETPOINT: await self.dhw_setpoint(),
+            SZ_DHW_TEMP: await self.dhw_temp(),
+            SZ_OEM_CODE: await self.oem_code(),
+            SZ_OUTSIDE_TEMP: await self.outside_temp(),
+            SZ_REL_MODULATION_LEVEL: await self.rel_modulation_level(),
+            SZ_CH_ACTIVE: await self.ch_active(),
+            SZ_CH_ENABLED: await self.ch_enabled(),
+            SZ_COOLING_ACTIVE: await self.cooling_active(),
+            SZ_COOLING_ENABLED: await self.cooling_enabled(),
+            SZ_DHW_ACTIVE: await self.dhw_active(),
+            SZ_DHW_BLOCKING: await self.dhw_blocking(),
+            SZ_DHW_ENABLED: await self.dhw_enabled(),
+            SZ_FAULT_PRESENT: await self.fault_present(),
+            SZ_FLAME_ACTIVE: await self.flame_active(),
+            SZ_SUMMER_MODE: await self.summer_mode(),
+            SZ_OTC_ACTIVE: await self.otc_active(),
+            #
+            # "status_opentherm": await self.opentherm_status(),
+            # "status_ramses_ii": await self.ramses_status(),
+        }
