+++ tests/tests_rf/test_api_schedule.py
@@ -7,7 +7,7 @@ from typing import Any
 import pytest
 
 from ramses_rf import Gateway
-from ramses_rf.device import Controller
+from ramses_rf.devices import Controller
 from ramses_rf.gateway import GatewayConfig
 from ramses_rf.systems import Evohome, Zone
 from ramses_rf.systems.schedule import InnerScheduleT
