+++ tests/utils/analyze_diff_rf.py
@@ -45,7 +45,7 @@ from typing import TYPE_CHECKING, Any, Final, cast
 from unittest.mock import AsyncMock, patch
 
 from ramses_rf import Gateway
-from ramses_rf.device import DeviceHeat, DeviceHvac
+from ramses_rf.devices import DeviceHeat, DeviceHvac
 from ramses_rf.gateway import GatewayConfig
 from ramses_tx.const import SZ_READER_TASK
 from ramses_tx.exceptions import TransportError
